Restructure code folders
