+++ application/src/AssetList/AssetListElement.js
