+++ application/src/DataRetrival/DataExtractor.js
